Moved ServiceMonitor to separate package
